requirements.txt
matplotlib==3.7.1
numpy==1.24.2
opencv_python==4.7.0.72
pandas==1.4.2
Pillow==9.5.0
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
torchmetrics==0.11.4
